requirements-docker.txt
asgiref==3.11.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dj-rest-auth==7.1.1
Django==5.2.11
django-allauth==65.14.3
django-cors-headers==4.9.0
django-environ==0.11.2
django-extensions==4.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
packaging==26.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.3.1
urllib3==2.0.3
gunicorn==21.2.0